pyproject.toml
[project]
name = "mamba-wikitext103"
version = "0.1.0"
description = "Training MAMBA-130M from scratch on WikiText-103"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.39.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "wandb>=0.16.0",
    "causal-conv1d>=1.4.0",
    "mamba-ssm>=2.2.0",
    "evaluate>=0.4.0",
    "tensorboard>=2.14.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
